[AIRFLOW-5769] Move the S3_hook to /providers/aws/hooks (#6443)
