+++ tests/test_core_to_contrib.py
@@ -132,6 +132,10 @@ HOOK = [
         "airflow.providers.aws.hooks.athena.AWSAthenaHook",
         "airflow.contrib.hooks.aws_athena_hook.AWSAthenaHook",
     ),
+    (
+        "airflow.providers.aws.hooks.s3.S3Hook",
+        "airflow.hooks.S3_hook.S3Hook",
+    ),
 ]
 OPERATOR = [
     (
